[X680x0] Fixed 1024x1024 graphic layer rendering, added basic interlaced mode support.
git-svn-id: svn://messdev.no-ip.org/mess@8879 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/includes/x68k.h
@@ -128,8 +128,8 @@ struct x68k_system
 		int hsync_end;  // Horizontal Sync End
 		int vsync_end;  // Vertical Sync End
 		int hsyncadjust;  // Horizontal Sync Adjustment
-		int hmultiple;  // Horizontal pixel multiplier
-		int vmultiple;  // Vertical scanline multiplier (x2 for doublescan modes)
+		float hmultiple;  // Horizontal pixel multiplier
+		float vmultiple;  // Vertical scanline multiplier (x2 for doublescan modes)
 		int height;
 		int width;
 		int visible_height;
